Add Day 9: Marble Mania
This took a little bit more effort on my side because of Elixir.
It seems that Double Linked Lists are a tough problem for immutable
data structures and I never thought about that before.

The way I chose to workaround that is how we design digraphs, with
a map of IDs pointing to the nodes, and the nodes reference each
other no directly but via these external IDs.

The first version using normal Elixir list was running the Part 1
input in ~30 seconds, with the new digraph version it's running
in 0.2 seconds (!!!!). I was not expecting this change on the
performance but this is just a reminder on how important is to
choose your right data abstractions.