Add Day 10: The Stars Align
This one is taking advantage of Stream.iterate to have an infinite
stream of points positions, the key aspect here is to calculate positions
for each new steps only when needed.

I struggled a little bit with the second part as I didn't want to hard code
the height of the text, to do that I'm calculating the bounding box of the
result and seeing if there's some vertical lines with all its positions filled
(like what happens with char I or T). It took me some time to realize
points could overlap, so I had to add a call to Enum.uniq before
counting for frequencies.