add solution for Day 11: Chronal Charge
I started this solution with a naive approach of doing all the sums all
the time, which worked for the first star but obviously didn't work
for the second one where you have to calculate all the possible
square sizes.

Then I tried to move to a version where I was caching the line sums and
calculating bigger squares combining smaller ones (for example
calculating the sum of 15 cells by looking up the cached sum for 5 cells
and combining them 5+5+5). The code grew complicated really fast and
the performance was still not enough for the second star.

Only then I realized that this problem could be solved by doing a rolling
window sum on both horizontal and vertical direction, after
implementing the code it passed the second star running in less then 5
secs. Elixir is quite good for this solution as doing Enum.zip on a 2d
array felt as a natural way on how to handle this kind of problems in real
life.