Add Day 12: Subterranean Sustainability
I build this solution thinking about a hasing function for the input, so
that I could use a rolling window sum and not have to read and compare
 5 chars for every new input. I decided to treat the input as a binary
input, with '#' as 1 and '.' as 0. This made the calculation of the next
window value just a bitwise operation, erasing the high bit, left shiftting
 the current content and adding the lower bit.

I'm taking advantage of Stream.unfold to generate new states on
demand. This was enough for the first part but it failed bad for the
second one. The funny part is that in Elixir even doing a `Stream.cycle`
 and dropping the first 5e10 elements was taking more then 1min, I'm
not sure if there's a faster way to do it.

From the beginning I decided to store the start index of the current
state as a field, this way I could trim initial zeroes and have a better data
structure. It proved to be useful to solve the second part, as I needed to
somehow find if the state was stable, which means that the binary data
was not going to change anymore, and do a "fast-forward" advance of
generations, not caring about the state itself but based on start position
deltas.