Add Day 13: Mine Cart Madness
The solution for this one was pretty straight forward, it was only
a matter of calculating the positions and following the track.

One thing that only became apparent after I started is that Y-axis
points downwards, which means I had to choose a definition for
south (points where Y is decreasing or points downwards on the
physical world). I decided to go with the definition of south is
where Y decreases and adapted the code to work like that.