Add Day 14: Chocolate Charts
This problem was super straight forward without a lot of problems
on the implementation. After solving past days problems with
streams now I feel very confident to use it and it seems to solve
the problem of not having a `while true` loop in Elixir.

Working with numbers in Elixir is made easy with methods like
digits, I used it in the first part to separate the generated score on
its digits.

While solving the second part I had to change my tick
implementation to always return a new item for every score
added, even if the tick added two scores (for example 1 and 0 for
10). After this it was just a matter of reducing the result with the
help of Stream.scan and doing some integer operations on the
score to transform the last N digits into a single number.

The solution for the second part took 3min to execute, maybe
there's some room for improvement on tick function but I doubt
that with a high number of iterations (>20m) there's any way to
get an exponential change.