Add Day 15: Beverage Bandits
Problem was much more about following the spec then
implementing difficult algorithms. The only tricky part was on how
to implement the path algorithm in Elixir, which turned out to not
be that dificult using Erlang queue.

The second part solution is just a brute force iteration on the first
part solution for different attack values, this could be improved by
stopping the game after the first elf dies.